Packages renaming and reordering.
