+++ src/main/java/com/proxiad/hangmangame/logic/GameSessionServiceImpl.java
@@ -1,6 +1,6 @@
-package com.proxiad.task.ivanboyukliev.hangmangame.service;
+package com.proxiad.hangmangame.logic;
 
-import static com.proxiad.task.ivanboyukliev.hangmangame.service.ApplicationConstants.INVALID_GAME_MSG;
+import static com.proxiad.hangmangame.util.ApplicationConstants.INVALID_GAME_MSG;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -8,6 +8,8 @@ import java.util.UUID;
 import javax.servlet.ServletException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.proxiad.hangmangame.model.GameSessionRepository;
+import com.proxiad.hangmangame.model.WordRepository;
 
 @Service
 public class GameSessionServiceImpl implements GameSessionService {
@@ -64,7 +66,6 @@ public class GameSessionServiceImpl implements GameSessionService {
 
     String wordToGuess = gameSession.getWordToGuess();
     long guessedLtrsOccurCnt = wordToGuess.chars().filter(ch -> ch == userInputLetter).count();
-
     String previousPuzzledWord = gameSession.getPuzzledWord();
 
     if (guessedLtrsOccurCnt == 0) {
