+++ src/main/java/com/proxiad/hangmangame/logic/InvalidGameSessionException.java
@@ -1,4 +1,4 @@
-package com.proxiad.task.ivanboyukliev.hangmangame.service;
+package com.proxiad.hangmangame.logic;
 
 import javax.servlet.ServletException;
