+++ src/main/java/com/proxiad/hangmangame/web/StartNewGameController.java
@@ -1,13 +1,13 @@
-package com.proxiad.task.ivanboyukliev.hangmangame.web;
+package com.proxiad.hangmangame.web;
 
-import static com.proxiad.task.ivanboyukliev.hangmangame.service.ApplicationConstants.GAME_BASE_URL;
+import static com.proxiad.hangmangame.util.ApplicationConstants.GAME_BASE_URL;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import com.proxiad.task.ivanboyukliev.hangmangame.service.GameSession;
-import com.proxiad.task.ivanboyukliev.hangmangame.service.GameSessionService;
+import com.proxiad.hangmangame.logic.GameSession;
+import com.proxiad.hangmangame.logic.GameSessionService;
 
 @Controller
 @RequestMapping(GAME_BASE_URL)
