+++ src/test/java/com/proxiad/task/ivanboyukliev/hangmangame/logic/GameSessionTest.java
@@ -1,7 +1,7 @@
-package com.proxiad.task.ivanboyukliev.hangmangame.service;
+package com.proxiad.task.ivanboyukliev.hangmangame.logic;
 
-import static com.proxiad.task.ivanboyukliev.hangmangame.service.ApplicationConstants.BONUS_TRIES;
-import static com.proxiad.task.ivanboyukliev.hangmangame.service.ApplicationConstants.SECRET_ENCODE_VAL;
+import static com.proxiad.hangmangame.util.ApplicationConstants.BONUS_TRIES;
+import static com.proxiad.hangmangame.util.ApplicationConstants.SECRET_ENCODE_VAL;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.ParameterizedTest.ARGUMENTS_PLACEHOLDER;
 import java.util.stream.Stream;
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import com.proxiad.hangmangame.logic.GameSession;
 
 class GameSessionTest {
