+++ src/test/java/it/web/StartNewGameControllerTest.java
@@ -1,6 +1,6 @@
 package it.web;
 
-import static com.proxiad.task.ivanboyukliev.hangmangame.service.ApplicationConstants.GAME_BASE_URL;
+import static com.proxiad.hangmangame.util.ApplicationConstants.GAME_BASE_URL;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
@@ -16,10 +16,10 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import com.proxiad.task.ivanboyukliev.hangmangame.WebConfiguration;
-import com.proxiad.task.ivanboyukliev.hangmangame.service.GameSession;
-import com.proxiad.task.ivanboyukliev.hangmangame.service.GameSessionService;
-import com.proxiad.task.ivanboyukliev.hangmangame.web.StartNewGameController;
+import com.proxiad.hangmangame.WebConfiguration;
+import com.proxiad.hangmangame.logic.GameSession;
+import com.proxiad.hangmangame.logic.GameSessionService;
+import com.proxiad.hangmangame.web.StartNewGameController;
 
 @ExtendWith(MockitoExtension.class)
 @ContextConfiguration(classes = WebConfiguration.class)
